Implement most of the new model merge strategy
We now check that when merging models that at least one in one of the
models all of the activities are part of the same process, and that
there are no activities that aren't part of that process.

Still need to fix duplicated chemicals and handle input/output when
comparing activities between models.

Refs pombase/pombase-gocam#13